feat:Finalizando busca por quantidade mínima

+++ src/main/java/autostock/api/autostock/useCases/product/DeleteProduct/DeleteProductUseCaseImpl.java
@@ -1,9 +1,11 @@
 package autostock.api.autostock.useCases.product.DeleteProduct;
 
 
-import org.springframework.stereotype.Component;
+import javax.validation.constraints.NotNull;
 
+import org.springframework.stereotype.Component;
 import autostock.api.autostock.repository.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 @Component
@@ -15,13 +17,19 @@ public class DeleteProductUseCaseImpl implements IDeleteProductUseCase {
         _productRepository=productRepository;
     
 }
+@SuppressWarnings("null")
 @Transactional
 public void deleteProduct(Long id){
-        try {
-          _productRepository.deleteById(id);
-
-        } catch (Exception e) {
-            throw new RuntimeException("Não é possível deletar");
+    try {
+        if (_productRepository.existsById(id)) {
+            _productRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Produto não encontrado com o ID fornecido: " + id);
         }
+    } catch (EntityNotFoundException e) {
+        throw e; 
+    } catch (Exception e) {
+        throw new RuntimeException("Não é possível deletar o produto com o ID: " + id, e);
+    }
 }
 }
